iot/esp/src/Core/Controller/Depth/Dream/requirements.txt
spherov2
bleak>=0.20.0
websocket-client
websockets
